refactor(engine): break up engine into multiple files
